pyproject.toml
[project]
name = "markspace"
version = "0.1.0"
description = "Stigmergic coordination protocol for LLM agent fleets"
license = "MIT AND CC-BY-4.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "httpx>=0.24,<1.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "hypothesis>=6.0",
]
experiments = [
    "matplotlib>=3.7",
    "numpy>=1.24",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.pyright]
include = ["markspace", "tests"]
pythonVersion = "3.11"
typeCheckingMode = "standard"

[dependency-groups]
dev = [
    "pre-commit (>=4.5.1,<5.0.0)",
    "scipy (>=1.17.1,<2.0.0)",
    "statsmodels (>=0.14.6,<0.15.0)"
]
